services/gateway/requirements.txt
fastapi
uvicorn[standard]
aio-pika
pydantic>=2
redis
aiohttp
pytest
pytest-asyncio
supabase
python-jose[cryptography]
